requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
autograd==1.8.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
diffusers==0.35.2
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.1
etils==1.13.0
exceptiongroup==1.3.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
immutabledict==4.2.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
keras==3.12.0
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mup==1.0.0
namex==0.1.0
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
openvino==2025.3.0
openvino-telemetry==2025.2.0
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
promise==2.3
protobuf==6.33.1
psutil==7.1.3
pyarrow==22.0.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-datasets==4.9.9
tensorflow-metadata==1.17.2
termcolor==3.2.0
tifffile==2025.5.10
toml==0.10.2
torch==2.9.1+cu126
torchvision==0.24.1+cu126
tqdm==4.67.1
triton-windows==3.5.1.post21
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.1
zipp==3.23.0